src/main/python/requirements.txt
cmake==3.27.7
colorama==0.4.6
colormap==1.0.6
colour==0.1.5
dlib==19.24.2
easydev==0.12.1
extcolors==1.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0

imageio==2.31.6
imageio-ffmpeg==0.4.9

nltk==3.8.1
numpy==1.23.5

opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78

pandas==1.5.3
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pandas-stubs==1.5.3.230304

Pillow==9.0.0
